refactor: Change the package structure (My Style)
