+++ src/main/java/com/hexagonal/courseregistration/user/adapter/persistence/JpaUserRepository.java
@@ -1,6 +1,6 @@
-package com.hexagonal.courseregistration.user.adapter.out.persistence;
+package com.hexagonal.courseregistration.user.adapter.persistence;
 
-import com.hexagonal.courseregistration.user.domain.Authority;
+import com.hexagonal.courseregistration.user.application.Authority;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
