+++ src/main/java/com/hexagonal/courseregistration/user/adapter/web/JsonRegisterRequest.java
@@ -1,6 +1,6 @@
-package com.hexagonal.courseregistration.user.adapter.in.web;
+package com.hexagonal.courseregistration.user.adapter.web;
 
-import com.hexagonal.courseregistration.user.domain.Authority;
+import com.hexagonal.courseregistration.user.application.Authority;
 import com.hexagonal.courseregistration.user.application.RegisterRequest;
 
 record JsonRegisterRequest(
