+++ src/main/java/com/hexagonal/courseregistration/user/application/AllUser.java
@@ -1,0 +1,7 @@
+package com.hexagonal.courseregistration.user.application;
+
+public interface AllUser {
+  boolean exist(String idNumber, Authority authority);
+
+  void register(User user);
+}
