+++ src/main/java/com/hexagonal/courseregistration/user/application/Authority.java
@@ -1,0 +1,5 @@
+package com.hexagonal.courseregistration.user.application;
+
+public enum Authority {
+  STUDENT, PROFESSOR
+}
