+++ src/main/java/com/hexagonal/courseregistration/user/application/RegisterRequest.java
@@ -1,7 +1,5 @@
 package com.hexagonal.courseregistration.user.application;
 
-import com.hexagonal.courseregistration.user.domain.Authority;
-
 public record RegisterRequest(
   String name,
   String idNumber,
